+++ src/runez/__init__.py
@@ -17,7 +17,7 @@ from runez.file import compress, copy, decompress, delete, filesize, ls_dir, mov
 from runez.logsetup import LogManager as log, ProgressBar
 from runez.program import check_pid, is_executable, make_executable, run, shell, which
 from runez.serialize import from_json, json_sanitized, read_json, represented_json, save_json, Serializable
-from runez.system import abort, abort_if, cached_property, uncolored, Undefined, UNSET, wcswidth
+from runez.system import abort, abort_if, cached_property, OptionalColor, uncolored, Undefined, UNSET, wcswidth
 from runez.system import Anchored, CaptureOutput, CurrentFolder, OverrideDryrun, TempArgv, TrackedOutput
 from runez.system import capped, decode, DEV, flattened, joined, quoted, resolved_path, short, stringified, SYS_INFO
 from runez.system import first_line, get_version, is_basetype, is_iterable, ltattr
@@ -38,7 +38,7 @@ __all__ = [  # noqa: RUF022, grouped and sorted by provenance module
     "log", "ProgressBar",
     "check_pid", "is_executable", "make_executable", "run", "shell", "which",
     "from_json", "json_sanitized", "read_json", "represented_json", "save_json", "Serializable",
-    "abort", "abort_if", "cached_property", "uncolored", "Undefined", "UNSET", "wcswidth",
+    "abort", "abort_if", "cached_property", "OptionalColor", "uncolored", "Undefined", "UNSET", "wcswidth",
     "Anchored", "CaptureOutput", "CurrentFolder", "OverrideDryrun", "TempArgv", "TrackedOutput",
     "capped", "decode", "DEV", "flattened", "joined", "quoted", "resolved_path", "short", "stringified", "SYS_INFO",
     "first_line", "get_version", "is_basetype", "is_iterable", "ltattr"